+++ src/main/java/br/inatel/icc/idp/artefato/model/ProductEntity.java
@@ -1,5 +1,6 @@
 package br.inatel.icc.idp.artefato.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,11 +21,12 @@ public class ProductEntity {
 
     private final String name;
     private final String description;
-    private final Double price;
+    private final BigDecimal price;
     private final Boolean isAvailable;
     private final List<String> imageURL;
 
-    public ProductEntity(String name, String description, Double price, Boolean isAvailable, List<String> imageURL) {
+    public ProductEntity(String name, String description, BigDecimal price, Boolean isAvailable,
+            List<String> imageURL) {
         this.id = null;
         this.name = name;
         this.description = description;
